+++ actix-protobuf/src/lib.rs
@@ -1,5 +1,6 @@
 #![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, nonstandard_style)]
+#![warn(future_incompatible)]
 
 use std::{
     fmt,
@@ -10,6 +11,7 @@ use std::{
 };
 
 use actix_web::{
+    body::BoxBody,
     dev::Payload,
     error::PayloadError,
     http::header::{CONTENT_LENGTH, CONTENT_TYPE},
@@ -144,6 +146,8 @@ where
 }
 
 impl<T: Message + Default> Responder for ProtoBuf<T> {
+    type Body = BoxBody;
+
     fn respond_to(self, _: &HttpRequest) -> HttpResponse {
         let mut buf = Vec::new();
         match self.0.encode(&mut buf) {