+++ actix-redis/CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
+* Update `actix-web` dependency to v4.0.0-beta.10. [#203]
 * Minimum supported Rust version (MSRV) is now 1.52.
+
+[#203]: https://github.com/actix/actix-extras/pull/203
 
 
 ## 0.10.0-beta.2 - 2020-06-27